refactor: merged enums.py into constants.py to reduce module sprawl for role and punishment definitions. Updated imports and usages.
