+++ moddingway/database/exiles_database.py
@@ -1,7 +1,7 @@
 import logging
 from datetime import datetime, timezone
 
-from moddingway.enums import ExileStatus
+from moddingway.constants import ExileStatus
 
 from . import DatabaseConnection
 from .models import Exile, PendingExile
