+++ moddingway/database/models/strike.py
@@ -3,7 +3,7 @@ from typing import Optional
 
 from pydantic import BaseModel
 
-from moddingway.enums import StrikeSeverity
+from moddingway.constants import StrikeSeverity
 from moddingway.settings import get_settings
 
 settings = get_settings()
