+++ moddingway/database/models/user.py
@@ -3,7 +3,7 @@ from typing import Optional
 
 from pydantic import BaseModel
 
-from moddingway.enums import UserRole
+from moddingway.constants import UserRole
 from moddingway.settings import get_settings
 
 settings = get_settings()
