+++ moddingway_api/routes/user_routes.py
@@ -5,7 +5,6 @@ from fastapi_pagination import Page
 
 from moddingway.database import users_database
 from moddingway_api.schemas.user_schema import User
-from moddingway.enums import UserRole
 from moddingway_api.utils.paginate import paginate, parse_pagination_params
 
 router = APIRouter(prefix="/users")
